+++ lib/cassandra/cluster/schema/fetchers.rb
@@ -1278,7 +1278,13 @@ module Cassandra
             name      = column_data['column_name']
             is_static = column_data['kind'].to_s.casecmp('STATIC').zero?
             order     = column_data['clustering_order'] == 'desc' ? :desc : :asc
-            type, is_frozen = @type_parser.parse(column_data['type'], types)
+            if column_data['type'][0] == "'"
+              # This is a custom column type.
+              type = Types.custom(column_data['type'].slice(1, column_data['type'].length - 2))
+              is_frozen = false
+            else
+              type, is_frozen = @type_parser.parse(column_data['type'], types)
+            end
 
             Column.new(name, type, order, is_static, is_frozen)
           end